Refactor internal/pinata package
